Added Controllers. Large amount of code moved to get rid of old core.php and pictsharemodel.php.
