+++ Classes/Autoloader.php
@@ -9,28 +9,6 @@ use PictShare\Classes\Exceptions\ClassNotFoundException;
 class Autoloader
 {
     /**
-     * Old autoloader, for backwards compatibility.
-     *
-     * @TODO: When everything is converted over to the namespace autoloader, remove this.
-     *
-     * @deprecated
-     *
-     * @param $className
-     *
-     * @return bool
-     */
-    public static function deprecatedAutoload(string $className): bool
-    {
-        if (file_exists(BASE_DIR . 'models/' . strtolower($className) . '.php')) {
-            include_once BASE_DIR . 'models/' . strtolower($className) . '.php';
-
-            return true;
-        }
-
-        return false;
-    }
-
-    /**
      * New autoloader, supporting namespaces. PSR-4 style.
      *
      * @param $className
@@ -64,20 +42,9 @@ class Autoloader
 
     /**
      * Temporary shim around the real autoloader.
-     *
-     * @TODO Drop the old loader completely.
      */
     public static function init()
     {
-        spl_autoload_register(
-            function ($className) {
-                // New style first.
-                $loaded = static::namespaceAutoloader($className);
-
-                if (!$loaded) {
-                    static::deprecatedAutoload($className);
-                }
-            }
-        );
+        spl_autoload_register('static::namespaceAutoloader');
     }
 }
