+++ Tests/bootstrap.php
@@ -11,3 +11,4 @@ require_once '../Classes/Autoloader.php';
 Autoloader::init();
 
 \define('BASE_DIR', __DIR__ . '/../');
+\define('UPLOAD_DIR', BASE_DIR . 'upload/');
