+++ composer.json
@@ -7,6 +7,8 @@
         "friendsofphp/php-cs-fixer": "^2.12",
         "squizlabs/php_codesniffer": "^3.3",
         "phan/phan": "^0.12.13",
-        "phpunit/phpunit": "^7.2"
+        "phpunit/phpunit": "^7.2",
+        "phpmd/phpmd": "^2.6",
+        "phpunit/php-code-coverage": "^6.0"
     }
 }
