+++ tools/backblaze_upload.php
@@ -3,6 +3,7 @@
 use PictShare\Classes\Autoloader;
 use PictShare\Classes\FileSizeFormatter;
 use PictShare\Classes\StorageProviderFactory;
+use PictShare\Models\BaseModel;
 
 /**
  * Backblaze uploader.
@@ -16,11 +17,10 @@ if (PHP_SAPI !== 'cli') {
 
 require_once '../Classes/Autoloader.php';
 require_once '../inc/config.inc.php';
-require_once '../inc/core.php';
 
 Autoloader::init();
 
-$pm         = new PictshareModel();
+$model      = new BaseModel();
 $sim        = false;
 $recentOnly = false;
 $uploadSize = 0;
@@ -55,7 +55,7 @@ while (false !== ($filename = readdir($dh))) {
     }
 
     $type = pathinfo($img, PATHINFO_EXTENSION);
-    $type = $pm->isTypeAllowed($type);
+    $type = $model->isTypeAllowed($type);
 
     if ($type) {
         if ($recentOnly === true) {
